Split htpy into multiple private modules
